Add rudimentary UI for adding books
