+++ app/build.gradle
@@ -60,9 +60,10 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.5.1'
     implementation 'androidx.activity:activity-compose:1.5.1'
     debugImplementation "androidx.compose.ui:ui-tooling:1.2.1"
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-compose:2.5.1'
     implementation "com.airbnb.android:lottie-compose:5.2.0"
     implementation 'androidx.navigation:navigation-compose:2.5.2'
+    implementation 'androidx.hilt:hilt-navigation-compose:1.0.0'
 
     implementation 'androidx.core:core-splashscreen:1.0.0'
