+++ app/src/main/java/dev/sanskar/nero/data/GoogleBooksResponse.kt
@@ -16,11 +16,11 @@ data class VolumeInfo(
     val authors: List<String>,
     val averageRating: Double = 0.0,
     val categories: List<String> = listOf(),
-    val description: String,
+    val description: String = "",
     val imageLinks: ImageLinks,
     val industryIdentifiers: List<IndustryIdentifier> = listOf(),
     val language: String,
-    val pageCount: Int,
+    val pageCount: Int = -1,
     val publishedDate: String,
     val publisher: String = "",
     val ratingsCount: Int = 0,
