+++ app/src/main/java/dev/sanskar/nero/ui/AddBook.kt
@@ -1,15 +1,64 @@
 package dev.sanskar.nero.ui
 
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Button
-import androidx.compose.material.Text
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.OutlinedTextField
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import dev.sanskar.nero.ui.components.BookRow
+import kotlinx.coroutines.delay
 
 @Composable
-fun AddBook(modifier: Modifier = Modifier, onClick: () -> Unit) {
-    Button(onClick = onClick, modifier = Modifier.padding(32.dp)) {
-        Text("Add Sample Data")
+fun AddBook(
+    modifier: Modifier = Modifier,
+    viewModel: MainViewModel = hiltViewModel(),
+) {
+    val focusRequester = FocusRequester()
+    Column(
+        modifier = modifier.fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        var query by remember { mutableStateOf("") }
+        OutlinedTextField(
+            value = query,
+            onValueChange = {
+                query = it
+                viewModel.searchGoogleBooks(it)
+            },
+            modifier = Modifier
+                .fillMaxWidth(0.9f)
+                .padding(32.dp)
+                .focusRequester(focusRequester)
+        )
+        LazyColumn {
+            items(viewModel.googleBooksSearchResult, key = { it.id }) {
+                BookRow(it) {
+                    viewModel.addBook(it)
+                }
+            }
+        }
+    }
+
+    LaunchedEffect(Unit) {
+        focusRequester.requestFocus()
     }
 }
